+++ kernel/numerical/matmul/C/OpenMP/MKL/double/dgemm_entry.c
@@ -36,8 +36,8 @@ int bi_entry(void *mcb, int problemSize,double *results){
 	double one=1.0;
 	double time=0, start, stop;
 	double nOperations=0.0;
-	long lCurrentSize;
-	unsigned long size;
+	long long lCurrentSize;
+	unsigned long long size;
 	char N='N';
 	double *f1, *f2, *f3;
 	int ii, jj;
@@ -46,7 +46,7 @@ int bi_entry(void *mcb, int problemSize,double *results){
 	if(results == NULL)
 		return -1;
 	
-	size = (unsigned long)bi_get_list_element(problemSize);
+	size = (unsigned long long)bi_get_list_element(problemSize);
 	results[0] = size;
 	nOperations = (1.0*size)*(1.0*size)*(2.0*size-1.0);
 	
@@ -97,5 +97,3 @@ int bi_entry(void *mcb, int problemSize,double *results){
 
 	return 0;
 }
-
-
